requirements.txt
asgiref==3.6.0
Django==4.2.1
django-ckeditor==6.0.0
mysqlclient==2.1.1
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
